RAYMOND added OwnerLanding UI
Added ownerlanding UI to link owner login and the owner page. It can see store, add product and go back. USERSTORY2

+++ app/src/main/java/com/example/b07project/Product.java
@@ -4,11 +4,11 @@ public class Product {
 
     String name;
     String brand;
-    Double price;
+    float price;
 
     public Product(){}
 
-    public Product(String name, String brand, Double price){
+    public Product(String name, String brand, float price){
         this.name = name;
         this.brand = brand;
         this.price = price;
